+++ src/codegen/machine/arch/x86_64/mod.rs
@@ -1,8 +1,12 @@
 use crate::codegen::machine::arch::TargetArchitecture;
+use crate::codegen::register::Register;
 use crate::{ast::ExprNode, codegen::CodeGenerationErr};
 
 /// X86_64 represents the x86_64 bit machine target.
 pub struct X86_64;
+
+mod register;
+use register::{GPRegisterAllocator, SizedGeneralPurpose};
 
 impl TargetArchitecture for X86_64 {}
 
@@ -21,92 +25,6 @@ impl SymbolTable {
             Some(())
         } else {
             None
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum SizedGeneralPurpose {
-    QuadWord(&'static str),
-    DoubleWord(&'static str),
-    Word(&'static str),
-    Byte(&'static str),
-}
-
-impl SizedGeneralPurpose {
-    /// returns the string representation of the register.
-    fn id(&self) -> &'static str {
-        match self {
-            SizedGeneralPurpose::QuadWord(id) => id,
-            SizedGeneralPurpose::DoubleWord(id) => id,
-            SizedGeneralPurpose::Word(id) => id,
-            SizedGeneralPurpose::Byte(id) => id,
-        }
-    }
-
-    fn operator_suffix(&self) -> &'static str {
-        match self {
-            SizedGeneralPurpose::QuadWord(_) => "q",
-            SizedGeneralPurpose::DoubleWord(_) => "l",
-            SizedGeneralPurpose::Word(_) => "w",
-            SizedGeneralPurpose::Byte(_) => "b",
-        }
-    }
-}
-
-impl std::fmt::Display for SizedGeneralPurpose {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let repr = self.id();
-        let register_suffix = match self {
-            SizedGeneralPurpose::QuadWord(_) => "",
-            SizedGeneralPurpose::DoubleWord(_) => "d",
-            SizedGeneralPurpose::Word(_) => "w",
-            SizedGeneralPurpose::Byte(_) => "b",
-        };
-
-        write!(f, "%{}{}", repr, register_suffix)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct GPRegisterAllocator {
-    registers: Vec<SizedGeneralPurpose>,
-}
-
-impl GPRegisterAllocator {
-    pub fn new(registers: Vec<SizedGeneralPurpose>) -> Self {
-        Self { registers }
-    }
-
-    /// Allocates a register for the duration of the life of closure.
-    fn allocate_then<F, R>(&mut self, f: F) -> R
-    where
-        F: FnOnce(&mut Self, &mut SizedGeneralPurpose) -> R,
-    {
-        self.registers
-            .pop()
-            .map(|mut reg| {
-                let ret_val = f(self, &mut reg);
-                self.registers.push(reg);
-                ret_val
-            })
-            .unwrap()
-    }
-}
-
-impl Default for GPRegisterAllocator {
-    fn default() -> Self {
-        Self {
-            registers: vec![
-                SizedGeneralPurpose::QuadWord("r8"),
-                SizedGeneralPurpose::QuadWord("r9"),
-                SizedGeneralPurpose::QuadWord("r10"),
-                SizedGeneralPurpose::QuadWord("r11"),
-                SizedGeneralPurpose::QuadWord("r12"),
-                SizedGeneralPurpose::QuadWord("r13"),
-                SizedGeneralPurpose::QuadWord("r14"),
-                SizedGeneralPurpose::QuadWord("r15"),
-            ],
         }
     }
 }
@@ -424,32 +342,4 @@ fn codegen_printint(reg: &mut SizedGeneralPurpose) -> Vec<String> {
         reg.operator_suffix(),
         reg
     )]
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::codegen::machine::arch::x86_64;
-
-    #[test]
-    fn should_allocate_a_register_from_an_unutilized_pool() {
-        assert_eq!(
-            ["r14", "r15"],
-            x86_64::GPRegisterAllocator::default().allocate_then(|allocator, reg| {
-                [allocator.allocate_then(|_, reg| reg.id()), reg.id()]
-            })
-        )
-    }
-
-    #[test]
-    fn should_free_allocations_on_scope_exit() {
-        let mut allocator = x86_64::GPRegisterAllocator::default();
-        let initial_len = allocator.registers.len();
-
-        // allocator pool should decrease by 1 while allocated in scope.
-        allocator
-            .allocate_then(|allocator, _| assert_eq!(initial_len - 1, allocator.registers.len()));
-
-        // register should be freed on scope exit.
-        assert_eq!(initial_len, allocator.registers.len());
-    }
 }
